Fix the Windows build.
git-svn-id: http://svn.powertabs.org/svn/ptedev/trunk@1176 5b1460fe-822e-4aa1-ac1b-8a6412c92c8c

+++ source/formats/guitar_pro/guitarproimporter.cpp
@@ -586,7 +586,8 @@ Position::DurationType GuitarProImporter::readDuration(Gp::InputStream &stream)
     // note, 1 -> eight note, etc
     // We need to convert to 1 = whole note, 2 = half note, 4 = quarter note,
     // etc
-    return static_cast<Position::DurationType>(std::pow(2.0, gpDuration + 2));
+    return static_cast<Position::DurationType>(
+        static_cast<int>(std::pow(2.0, gpDuration + 2)));
 }
 
 void GuitarProImporter::readChordDiagram(Gp::InputStream &stream,
